sprint 14 1

+++ controllers/cards.js
@@ -27,14 +27,19 @@ module.exports.createCard = (req, res) => {
 }; 
 
 module.exports.deleteCardById = (req, res) => {
-
-  Card.findByIdAndRemove(req.params.cardId,{ new: true })
-    .then(card => {
-      if(!card) {
-        res.status(404).send({ message: 'Карточки с таким Id не существует' })
-      } 
-      res.send({ data: card })
-    })
+  Card.findById(req.params.cardId)
+  .then((card)=>{
+    if(card.owner.toString()!==req.user._id){
+      res.status(400).send({ message: 'Нелья удалить чужую карточку' })
+    } else {
+      Card.findByIdAndRemove(req.params.cardId,{ new: true })
+      .then(card => {
+        if(!card) {
+          res.status(404).send({ message: 'Карточки с таким Id не существует' })
+        } 
+        res.send({ data: card })
+      })
+    }})
     .catch((err) => {
       if (err instanceof mongoose.Error.CastError) {
         res.status(400).send({ message: 'Карточки с таким Id не существует' })
